+++ lib/flutter_barrage.dart
@@ -17,7 +17,7 @@ class BarrageWall extends StatefulWidget {
   final int speed;
 
   /// used to adjust speed for each channel
-  final int speedCorrection;
+  final int speedCorrectionInMilliseconds;
 
   final double width;
   final double height;
@@ -40,7 +40,7 @@ class BarrageWall extends StatefulWidget {
     List<Bullet> bullets,
     BarrageWallController controller,
     ValueNotifier<BarrageValue> timelineNotifier,
-    this.speed,
+    this.speed = 5,
     this.child,
     this.width,
     this.height,
@@ -48,17 +48,14 @@ class BarrageWall extends StatefulWidget {
     this.maxBulletHeight,
     this.debug = false,
     this.safeBottomHeight = 0,
-    this.speedCorrection = 0,
+    this.speedCorrectionInMilliseconds = 3000,
   })  : controller = controller ??
-            BarrageWallController.withBarrages(bullets,
-                timelineNotifier: timelineNotifier),
+            BarrageWallController.withBarrages(bullets, timelineNotifier: timelineNotifier),
         selfCreatedController = controller == null {
     if (controller != null) {
-      this.controller.value = controller.value.size == 0
-          ? BarrageWallValue.fromList(bullets ?? [])
-          : controller.value;
-      this.controller.timelineNotifier =
-          controller.timelineNotifier ?? timelineNotifier;
+      this.controller.value =
+          controller.value.size == 0 ? BarrageWallValue.fromList(bullets ?? []) : controller.value;
+      this.controller.timelineNotifier = controller.timelineNotifier ?? timelineNotifier;
     }
   }
 
@@ -68,34 +65,30 @@ class BarrageWall extends StatefulWidget {
 
 class BulletPos {
   int id;
-  double position;
+  int channel;
+  double position; // from right to left
   double width;
   bool released = false;
   int lifetime;
+  Widget widget;
 
-  BulletPos({this.id, this.position, this.width})
+  BulletPos({this.id, this.channel, this.position, this.width, this.widget})
       : lifetime = DateTime.now().millisecondsSinceEpoch;
 
-  updateWith({int id, double position, double width = 0}) {
-    if (id == this.id) {
-      this.position = position;
-      this.width = width > 0 ? width : this.width;
-    } else {
-      if (this.position > position) {
-        this.id = id;
-        this.position = position;
-        this.width = width > 0 ? width : this.width;
-        this.released = false;
-        this.lifetime = DateTime.now().millisecondsSinceEpoch;
-      }
-    }
+  updateWith({double position, double width = 0}) {
+    this.position = position;
+    this.width = width > 0 ? width : this.width;
+    this.lifetime = DateTime.now().millisecondsSinceEpoch;
+//    debugPrint('update to $this');
   }
 
-  bool get hasExtraSpace => position > width + 8;
+  bool get hasExtraSpace {
+    return position > width + 8;
+  }
 
   @override
   String toString() {
-    return 'BulletPos{id: $id, position: $position, width: $width, released: $released}';
+    return 'BulletPos{id: $id, channel: $channel, position: $position, width: $width, released: $released, widget: $widget}';
   }
 }
 
@@ -117,8 +110,7 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
 
   int _calcSafeHeight(double height) => height.isInfinite
       ? context.size.height.toInt()
-      : (height - (_controller.safeBottomHeight ?? widget.safeBottomHeight))
-          .toInt();
+      : (height - (_controller.safeBottomHeight ?? widget.safeBottomHeight)).toInt();
 
   /// null means no available channels exists
   int _nextChannel() {
@@ -132,8 +124,7 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
     var channel = _random.nextInt(_randomSeed);
     var channelCode = 1 << channel;
 
-    while (
-        _usedChannel & channelCode != 0 && _usedChannel ^ _channelMask != 0) {
+    while (_usedChannel & channelCode != 0 && _usedChannel ^ _channelMask != 0) {
       times++;
       channel = channel >= _totalChannels ? 0 : channel + 1;
       channelCode = 1 << channel;
@@ -151,12 +142,13 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
   }
 
   _releaseChannels() {
+//    final now = DateTime.now().millisecondsSinceEpoch;
     for (int i = 0; i < _lastBullets.length; i++) {
       final channel = _lastBullets.keys.elementAt(i);
       var isNotReleased = !_lastBullets[channel].released;
-      var liveTooLong = false; // now - _lastBullets[channel].lifetime > 3000;
-      if (liveTooLong ||
-          (isNotReleased && _lastBullets[channel].hasExtraSpace)) {
+      var liveTooLong =
+          false; // now - _lastBullets[channel].lifetime > widget.speed * 2 * 1000 + widget.speedCorrectionInMilliseconds;
+      if (liveTooLong || (isNotReleased && _lastBullets[channel].hasExtraSpace)) {
         _lastBullets[channel].released = true;
         _usedChannel &= _channelMask ^ 1 << channel;
       }
@@ -172,10 +164,10 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
     // cannot get the width of widget when not rendered, make a twice longer width for now
     end ??= width * 2;
 
+    _releaseChannels();
     bullets.forEach((Bullet bullet) {
-      AnimationController controller;
+      AnimationController animationController;
 
-      _releaseChannels();
       final nextChannel = _nextChannel();
       if (nextChannel != null) {}
 
@@ -184,13 +176,12 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
         return;
       }
 
-      final showTimeInMilliseconds = (widget.speed ?? 5) * 2 * 1000 -
-          _speedCorrectionForChannels[nextChannel];
-      controller = AnimationController(
-          duration: Duration(milliseconds: showTimeInMilliseconds),
-          vsync: this);
+      final showTimeInMilliseconds =
+          widget.speed * 2 * 1000 - _speedCorrectionForChannels[nextChannel];
+      animationController = AnimationController(
+          duration: Duration(milliseconds: showTimeInMilliseconds), vsync: this);
       Animation<double> animation =
-          new Tween<double>(begin: 0, end: end).animate(controller..forward());
+          new Tween<double>(begin: 0, end: end).animate(animationController..forward());
 
       final channelHeightPos = nextChannel * _maxBulletHeight;
 
@@ -200,11 +191,12 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
         animation: animation,
         child: bullet.child,
         builder: (BuildContext context, Widget child) {
-          var widgetWidth = 0.0;
           if (animation.isCompleted) {
+            _lastBullets[nextChannel]?.updateWith(position: double.infinity);
             return const SizedBox();
           }
 
+          var widgetWidth = 0.0;
           if (context.findRenderObject() != null) {
             final RenderBox box = context.findRenderObject();
 
@@ -215,33 +207,37 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
             if (box != null &&
                 RenderObject.debugActiveLayout == null &&
                 animation.value > (fixedWidth + widgetWidth)) {
+              _lastBullets[nextChannel]?.updateWith(position: double.infinity);
               return const SizedBox();
             }
           }
 
-          final widthPos = fixedWidth - animation.value;
-
-          _releaseChannels();
-          if (!_lastBullets.containsKey(nextChannel)) {
+//          debugPrint(
+//              '${_lastBullets[nextChannel]?.id} == ${context.hashCode} $child pos: ${animation.value}');
+          // 【通道不为空】或者【通道的最后元素】之后出现了可以新增的元素
+          if (!_lastBullets.containsKey(nextChannel) ||
+              (_lastBullets.containsKey(nextChannel) &&
+                  _lastBullets[nextChannel].position > animation.value)) {
             _lastBullets[nextChannel] = BulletPos(
                 id: context.hashCode,
+                channel: nextChannel,
                 position: animation.value,
-                width: widgetWidth);
-          }
+                width: widgetWidth,
+                widget: child);
+//            debugPrint('add ${_lastBullets[nextChannel]} - ${context.hashCode}');
+          } else if (_lastBullets[nextChannel].id == context.hashCode) {
+            // 当前元素是最后元素，更新相关信息
+            _lastBullets[nextChannel]?.updateWith(position: animation.value, width: widgetWidth);
+          } // 其他情况直接更新页面元素
 
-          _lastBullets[nextChannel]?.updateWith(
-            id: context.hashCode,
-            position: animation.value,
-            width: widgetWidth,
-          );
-
+          final widthPos = fixedWidth - animation.value;
           return Transform.translate(
             offset: Offset(widthPos, channelHeightPos.toDouble()),
             child: child,
           );
         },
       );
-      _widgets.putIfAbsent(controller, () => bulletWidget);
+      _widgets.putIfAbsent(animationController, () => bulletWidget);
     });
   }
 
@@ -256,7 +252,7 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
   }
 
   void handleBullets() {
-    if (_processed != _controller.value.processedSize) {
+    if (_controller.isEnabled && _processed != _controller.value.processedSize) {
       if (_width == null || _height == null) {
         return;
       }
@@ -267,8 +263,8 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
         _channelMask = (2 << _totalChannels) - 1;
 
         List<int>.generate(_totalChannels + 1, (i) {
-          _speedCorrectionForChannels.add(widget.speedCorrection > 0
-              ? _random.nextInt(widget.speedCorrection)
+          _speedCorrectionForChannels.add(widget.speedCorrectionInMilliseconds > 0
+              ? _random.nextInt(widget.speedCorrectionInMilliseconds)
               : 0);
         });
       }
@@ -289,9 +285,11 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
     _controller.initialize();
 
     _controller.addListener(handleBullets);
+    _controller.enabledNotifier.addListener(() {
+      setState(() {});
+    });
 
-    /*
-    _cleaner = Timer.periodic(Duration(seconds: 10), (timer) {
+    _cleaner = Timer.periodic(Duration(milliseconds: 100), (timer) {
       _widgets.removeWhere((controller, widget) {
         if (controller.isCompleted) {
           controller.dispose();
@@ -299,7 +297,7 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
         }
         return false;
       });
-    });*/
+    });
 
     super.initState();
   }
@@ -327,34 +325,31 @@ class _BarrageState extends State<BarrageWall> with TickerProviderStateMixin {
         debugPrint("Timeline: ${_controller.timeline}");
         debugPrint("Bullets: ${_widgets.length}");
         debugPrint("UsedChannels: ${_usedChannel.toRadixString(2)}");
+        debugPrint("LastBullets[0]: ${_lastBullets[0]}");
       }
-      return Stack(
-        fit: StackFit.expand,
-        children: <Widget>[
-          widget.debug
-              ? Container(
-                  color: Colors.lightBlueAccent.withOpacity(0.7),
-                  child: Column(
+      return Stack(fit: StackFit.expand, children: <Widget>[
+        widget.debug
+            ? Container(
+                color: Colors.lightBlueAccent.withOpacity(0.7),
+                child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     mainAxisAlignment: MainAxisAlignment.end,
                     children: <Widget>[
                       Text('BarrageWallValue: ${_controller.value}'),
-                      Text(
-                          'TimelineNotifier: ${_controller.timelineNotifier?.value}'),
+                      Text('TimelineNotifier: ${_controller.timelineNotifier?.value}'),
                       Text('Timeline: ${_controller.timeline}'),
                       Text('Bullets: ${_widgets.length}'),
                       Text('UsedChannels: ${_usedChannel.toRadixString(2)}'),
-                    ],
-                  ),
-                )
-              : const SizedBox(),
-          widget.child,
-          Stack(
-            fit: StackFit.loose,
-            children: <Widget>[]..addAll(_widgets.values ?? const SizedBox()),
-          ),
-        ],
-      );
+                      Text('LastBullets[0]: ${_lastBullets[0]}'),
+                    ]))
+            : const SizedBox(),
+        widget.child,
+        _controller.isEnabled
+            ? Stack(
+                fit: StackFit.loose,
+                children: <Widget>[]..addAll(_widgets.values ?? const SizedBox()))
+            : const SizedBox(),
+      ]);
     });
   }
 }
@@ -375,8 +370,7 @@ class HashList<T> {
       if (_map.containsKey(key)) {
         _map[key].add(value);
       } else {
-        _map.putIfAbsent(
-            key, () => TreeSet<T>(comparator: comparator)..add(value));
+        _map.putIfAbsent(key, () => TreeSet<T>(comparator: comparator)..add(value));
       }
     });
   }
@@ -393,9 +387,8 @@ class BarrageValue {
 
   BarrageValue({this.timeline = -1, this.isPlaying = false});
 
-  BarrageValue copyWith({int timeline, bool isPlaying}) => BarrageValue(
-      timeline: timeline ?? this.timeline,
-      isPlaying: isPlaying ?? this.isPlaying);
+  BarrageValue copyWith({int timeline, bool isPlaying}) =>
+      BarrageValue(timeline: timeline ?? this.timeline, isPlaying: isPlaying ?? this.isPlaying);
 
   @override
   String toString() {
@@ -412,9 +405,9 @@ class BarrageWallValue {
 
   BarrageWallValue.fromList(List<Bullet> bullets,
       {this.showedTimeBefore = 0, this.waitingList = const []})
-      : bullets = HashList<Bullet>(
-            keyCalculator: (t) => Duration(milliseconds: t.showTime).inMinutes)
-          ..appendByMinutes(bullets),
+      : bullets =
+            HashList<Bullet>(keyCalculator: (t) => Duration(milliseconds: t.showTime).inMinutes)
+              ..appendByMinutes(bullets),
         size = bullets.length,
         processedSize = 0;
 
@@ -448,24 +441,25 @@ class BarrageWallValue {
 
 class BarrageWallController extends ValueNotifier<BarrageWallValue> {
   ValueNotifier<BarrageValue> timelineNotifier;
-  Timer _timer;
-  bool _isDisposed = false;
   int timeline = 0;
   int safeBottomHeight;
+  ValueNotifier<bool> enabledNotifier = ValueNotifier(true);
+  Timer _timer;
+  bool _isDisposed = false;
+
+  get isEnabled => enabledNotifier.value;
 
   BarrageWallController({List<Bullet> bullets, this.timelineNotifier})
       : super(BarrageWallValue.fromList(bullets ?? const []));
 
-  BarrageWallController.withBarrages(List<Bullet> bullets,
-      {this.timelineNotifier})
+  BarrageWallController.withBarrages(List<Bullet> bullets, {this.timelineNotifier})
       : super(BarrageWallValue.fromList(bullets ?? const []));
 
   Future<void> initialize() async {
     final Completer<void> initializingCompleter = Completer<void>();
 
     if (timelineNotifier == null) {
-      _timer = Timer.periodic(const Duration(milliseconds: 100),
-          (Timer timer) async {
+      _timer = Timer.periodic(const Duration(milliseconds: 100), (Timer timer) async {
         if (_isDisposed) {
           timer.cancel();
           return;
@@ -500,8 +494,7 @@ class BarrageWallController extends ValueNotifier<BarrageWallValue> {
     if (exists || bullets.isNotEmpty) {
       List<Bullet> toBePrecessed = value.bullets._map[key]
               ?.where((barrage) =>
-                  barrage.showTime > value.showedTimeBefore &&
-                  barrage.showTime <= timeline)
+                  barrage.showTime > value.showedTimeBefore && barrage.showTime <= timeline)
               ?.toList() ??
           [];
 
@@ -514,6 +507,16 @@ class BarrageWallController extends ValueNotifier<BarrageWallValue> {
     }
   }
 
+  disable() {
+    debugPrint("disable barrage ... current: $enabledNotifier");
+    enabledNotifier.value = false;
+  }
+
+  enable() {
+    debugPrint("enable barrage ... current: $enabledNotifier");
+    enabledNotifier.value = true;
+  }
+
   send(List<Bullet> bullets) {
     tryFire(bullets: bullets);
   }
@@ -524,7 +527,8 @@ class BarrageWallController extends ValueNotifier<BarrageWallValue> {
       _timer?.cancel();
     }
     _isDisposed = true;
-    timelineNotifier?.removeListener(_handleTimelineNotifier);
+    timelineNotifier?.dispose();
+    enabledNotifier.dispose();
     super.dispose();
   }
 }